--- apps/web-client/src/worldModels/index.d.ts
@@ -1,7 +1,0 @@
-declare module "worker-loader!*" {
-  class WebpackWorker extends Worker {
-    constructor();
-  }
-
-  export default WebpackWorker;
-}
